% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji-extraction.R
\name{emoji_extract_nest}
\alias{emoji_extract_nest}
\title{Emoji extraction nested summary}
\usage{
emoji_extract_nest(tweet_tbl, tweet_text)
}
\arguments{
\item{tweet_tbl}{A dataframe/tibble containing tweets/text.}

\item{tweet_text}{The tweet/text column.}
}
\value{
The original dataframe/tibble with an extra column collumn called
\code{.emoji_unicode}.
}
\description{
This function adds an extra list column called \code{.emoji_unicode} to the
original data, with all Emojis included.
}
