% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji-extraction.R
\name{emoji_extract_unnest}
\alias{emoji_extract_unnest}
\title{Emoji extraction unnested summary}
\usage{
emoji_extract_unnest(tweet_tbl, tweet_text)
}
\arguments{
\item{tweet_tbl}{A dataframe/tibble containing tweets/text.}

\item{tweet_text}{The tweet/text column.}
}
\value{
A summary tibble with the original row number and Emoji count.
}
\description{
If users would like to know how many Emojis and what kinds of Emojis each
Tweet has, \code{emoji_extract} is a useful function to output a global
summary with the row number of each Tweet containing Emoji and the Unicodes
associated with each Tweet.
}
