% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji-summary.R
\name{emoji_summary}
\alias{emoji_summary}
\title{Emoji Summary Tibble}
\usage{
emoji_summary(tweet_tbl, tweet_text)
}
\arguments{
\item{tweet_tbl}{A dataframe/tibble containing tweets.}

\item{tweet_text}{The column that is the tweet column.}
}
\value{
A summary tibble including # of tweets in total and # of tweets that
have at least one Emoji.
}
\description{
Emoji Summary Tibble
}
