% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji-summary.R
\name{emoji_tweets}
\alias{emoji_tweets}
\title{Emoji Text/Tweets Output}
\usage{
emoji_tweets(tweet_tbl, tweet_text)
}
\arguments{
\item{tweet_tbl}{A dataframe/tibble containing tweets/text.}

\item{tweet_text}{The tweet/text column.}
}
\value{
A dataframe/tibble containing only text with at least one Emoji
}
\description{
When users just want to focus on Tweets containing Emoji(s),
\code{emoji_tweets} filters out non-Emoji rows and only returns rows that
have at least one Emoji.
}
