% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoji_unicode_crosswalk}
\alias{emoji_unicode_crosswalk}
\title{Emoji name, Unicode, and Emoji category crosswalk}
\format{
A data frame with 4536 rows and 3 columns:
\describe{
  \item{emoji_name}{The name of Emoji per se.}
  \item{unicode}{The Unicode of Emoji.}
  \item{emoji_category}{The category Emoji falls into.}
}
}
\source{
The raw data sets (\code{emoji_name} and \code{emojis}) come from the
\code{emoji} package, and they are processed by the author for the specific
needs of \code{tidyEmoji}.
}
\usage{
emoji_unicode_crosswalk
}
\description{
A data set containing each Emoji name (such as grinning, smile), its
respective Unicode and category. One thing to note here is there are
duplicated Unicodes in the data set, because one Unicode could have multiple
Emoji names.
}
\keyword{datasets}
